refactor(api): add schema suffix to module validation Joi files
