+++ api/scripts/modulix/convert-joi-schema-to-json-schema.js
@@ -1,5 +1,5 @@
 import { convertJoiToJsonSchema } from '../../src/devcomp/infrastructure/datasources/conversion/joi-to-json-schema.js';
-import { moduleSchema } from '../../tests/devcomp/unit/infrastructure/datasources/learning-content/validation/module.js';
+import { moduleSchema } from '../../tests/devcomp/unit/infrastructure/datasources/learning-content/validation/module-schema.js';
 
 const jsonSchema = JSON.stringify(convertJoiToJsonSchema(moduleSchema), null, 2);
