+++ api/tests/devcomp/unit/infrastructure/datasources/learning-content/module-datasource_test.js
@@ -2,7 +2,7 @@ import moduleDatasource from '../../../../../../src/devcomp/infrastructure/datas
 import { LearningContentResourceNotFound } from '../../../../../../src/shared/infrastructure/datasources/learning-content/LearningContentResourceNotFound.js';
 import { expect } from '../../../../../test-helper.js';
 import { joiErrorParser } from './validation/joi-error-parser.js';
-import { moduleSchema } from './validation/module.js';
+import { moduleSchema } from './validation/module-schema.js';
 
 const modules = await moduleDatasource.list();
