+++ api/tests/devcomp/unit/infrastructure/datasources/learning-content/validation/element/image-schema.js
