+++ api/tests/devcomp/unit/infrastructure/datasources/learning-content/validation/element/qcm-schema.js
